requirements.txt
Appium-Python-Client==1.3.0
selenium==3.141.0
urllib3==1.26.20
pytest>=8.0.0
pytest-html>=4.0.0
fpdf2>=2.7.0
Pillow>=11.0.0
